Major styling changes to asl.css that mostly apply to forum_post_view.php and the vlog section. New backgrounds for post titles and new location on the page for the post message tools (edit, delete). This applies to reply tools (edit, delete) as well.
Major changes to forum_post_view.php in terms of where post editing tools are displayed (location on the page/location inside the post or reply row) and the visual styling of the post title/message. 

Major changes to vlog section through vlog_entry_view.php. 
Vlogs have identical styling as forum posts and replies. This includes displaying the vlog_entry title and message together with the user's avatar as one main element at the top of the page, as well as the vlog comments that follow beneath.

Code changes to forums.inc.php includes creation of new methods to support easier generation of the replies table.
Code changes to vlogs.inc.php that mirror the changes in forums.inc.php, but as applied to the vlog section and vlog tables in database.