Lots of changes to asl.css applied to top level post display pages such as forum_posts.php. Removed in-line styles wherever possible in the PHP code (to be followed by subsequent commit) to provide a finer level of specificity and control when displaying post titles.
Creation of classes to replace classless div's and anonymous in-line styled elements in PHP.  Created a div.title-inner and div.title-upper to be used for all post title thumbnails as previous code used anonymous, classless divs. Changed size and weight of post thumbnail text that would list the last activity and other elements about the post.  Restored the display of the hand icon underneath a post on display pages for main groups, but this time using the background-image css property.  

Created a class to add subtle 'glowish' effect on post objects for top level display pages. Hopefully this will make the post object more unified, as previously only the top 1/3rd was highlighted when the user mouses over it.